Правки module3-task1
